Added the initial files to describe the different core features of Photonic
CommunicationPrimitive is something similar to the Message object from the
Message library but this is way more generic;

PhotonicInitiator: This class is used as a bridge between the two application
domains. The initiators act as borders to each respective AppDomain. The beauty
of this design is that both Application domains will contain the Photonic
library so it's possible to have communication safely occur between the
application domains without having libraries specific to one of them being
loaded in the other.